Changed allow_outside_tests feature behavior
Previously, the feature attempted to conditionally generate an output
of an empty token stream if the feature wasn't enabled. This worked,
however the macro wouldn't work correctly when running tests without
the feature. Instead, the macro now generates structs/impls with the
cfg checks above them to let the next step in the compilation process
handle whether or not to actually include the generated code based
on whether tests are running or the allow_outside_tests feature is
enabled.