Fix: allow "index" field; Added documentation
Previously, the inner struct was setup to match fields with
the source struct. This meant that any fields that existed
on the inner struct would collide with those on the source
struct. This change adds an "inner_" prefix to avoid collisons.
The only field that currently existed was "index" which
now is available to use on source structs along with any
future fields that might be added.